ctrlsw: update to revision d20220317
Acked-for-series:Varunkumar Allagadapa <varunkumar.allagadapa@xilinx.com>

+++ lib_common/LevelLimit.h
@@ -1,6 +1,6 @@
 /******************************************************************************
 *
-* Copyright (C) 2008-2020 Allegro DVT2.  All rights reserved.
+* Copyright (C) 2008-2022 Allegro DVT2.  All rights reserved.
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
@@ -41,7 +41,7 @@
 /*************************************************************************/
 typedef struct AL_t_LevelLimit
 {
-  uint32_t uLimit;
+  uint64_t uLimit;
   uint8_t uLevel;
 }AL_TLevelLimit;
